+++ tests/model/model.cpp
@@ -164,11 +164,11 @@ TEST(Model, countComponents) {
     c1->setName("child1");
     c2->setName("child2");
 
-    EXPECT_EQ(0u, m.componentCount());
+    EXPECT_EQ(size_t(0), m.componentCount());
 
     m.addComponent(c1);
     m.addComponent(c2);
-    EXPECT_EQ(2u, m.componentCount());
+    EXPECT_EQ(size_t(2), m.componentCount());
 }
 
 TEST(Model, containsComponent) {
@@ -207,9 +207,9 @@ TEST(Model, removeComponent) {
     m.addComponent(c1);
     m.addComponent(c2);
 
-    EXPECT_EQ(2u, m.componentCount());
+    EXPECT_EQ(size_t(2), m.componentCount());
     EXPECT_TRUE(m.removeComponent(0));
-    EXPECT_EQ(1u, m.componentCount());
+    EXPECT_EQ(size_t(1), m.componentCount());
 
     libcellml::Printer printer;
     std::string a = printer.printModel(m);
@@ -221,13 +221,13 @@ TEST(Model, removeComponent) {
 
     // Remove the first occurence of "child1".
     EXPECT_TRUE(m.removeComponent("child1"));
-    EXPECT_EQ(2u, m.componentCount());
+    EXPECT_EQ(size_t(2), m.componentCount());
     a = printer.printModel(m);
     EXPECT_EQ(e2, a);
 
     // Expect no change to model.
     EXPECT_FALSE(m.removeComponent("child3"));
-    EXPECT_EQ(2u, m.componentCount());
+    EXPECT_EQ(size_t(2), m.componentCount());
 }
 
 TEST(Model, getComponentMethods) {
@@ -272,7 +272,7 @@ TEST(Model, takeComponentMethods) {
     m.addComponent(c2);
 
     libcellml::ComponentPtr c02 = m.takeComponent(1);
-    EXPECT_EQ(1u, m.componentCount());
+    EXPECT_EQ(size_t(1), m.componentCount());
 
     EXPECT_EQ(m.takeComponent(4), nullptr);
 
@@ -280,7 +280,7 @@ TEST(Model, takeComponentMethods) {
 
     libcellml::ComponentPtr c01 = m.takeComponent("child1");
     EXPECT_NE(nullptr, c01);
-    EXPECT_EQ(0u, m.componentCount());
+    EXPECT_EQ(size_t(0), m.componentCount());
 
     EXPECT_EQ("child1", c01->getName());
 
@@ -289,9 +289,9 @@ TEST(Model, takeComponentMethods) {
     EXPECT_EQ(e, a);
 
     // Expect no change.
-    EXPECT_EQ(0u, m.componentCount());
+    EXPECT_EQ(size_t(0), m.componentCount());
     EXPECT_EQ(nullptr, m.takeComponent("child4"));
-    EXPECT_EQ(0u, m.componentCount());
+    EXPECT_EQ(size_t(0), m.componentCount());
 }
 
 static int count = 0;
